Create question from parent CRUD of question Fix of redirection for createfromparent Statistic Converters

+++ Thesis JBM/src/main/java/services/QuestionService.java
@@ -79,21 +79,21 @@ public class QuestionService {
 	 */
 	public void delete(Question question) {
 		Assert.notNull(question);
-
-		for (Exercise e: question.getExercises()){
-			e.getQuestions().remove(question);				
+		for (Exercise e : question.getExercises()) {
+			e.getQuestions().remove(question);
 		}
-		
-		for (Metadata m: question.getMetadata()){
-			m.getQuestions().remove(question);				
-		}
-		
-		question.setExercises(new ArrayList<Exercise>());
-		question.setMetadata(new ArrayList<Metadata>());
-		
-		
 
-		questionRepository.delete(question);
+		for (Metadata m : question.getMetadata()) {
+			m.getQuestions().remove(question);
+		}
+
+		if (question.getExercises().size() == 0) {
+			question.setExercises(new ArrayList<Exercise>());
+			question.setMetadata(new ArrayList<Metadata>());
+			questionRepository.delete(question);
+		} else {
+			questionRepository.save(question);
+		}
 	}
 
 	public Collection<Question> findByExerciseId(int exerciseId) {
